src/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyYAML
openai
psycopg2-binary
python-json-logger